fix react context file
