+++ src/styles/globals.css
@@ -14,6 +14,3 @@ a {
 * {
   box-sizing: border-box;
 }
-
-
-
